feat: split input.js, output.js
