+++ service/movieApi/src/input/init.js
@@ -1,0 +1,7 @@
+export const mod = {}
+export default {}
+
+export const init = ({ fs }) => {
+  mod.fs = fs
+}
+
