+++ service/movieApi/src/output/makeDir.js
@@ -1,0 +1,7 @@
+import { mod } from './init.js'
+export default {}
+
+export const makeDir = ({ dirPath }) => {
+  return mod.fs.mkdirSync(dirPath, { recursive: true })
+}
+
