+++ service/movieApi/src/output/saveFile.js
@@ -1,0 +1,7 @@
+import { mod } from './init.js'
+export default {}
+
+export const saveFile = ({ filePath, fileBuffer }) => {
+  return mod.fs.writeFileSync(filePath, Buffer.from(fileBuffer))
+}
+
